Split vector implementations by family
